reorganize shell state control into a common separate assembly
