+++ src/AvaloniaIDE.Shell.Abstractions/AvaloniaIDE.Shell.Abstractions.csproj
@@ -1,0 +1,8 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <AssemblyName>AvaloniaIDE.Shell.Abstractions</AssemblyName>
+    <RootNamespace>AvaloniaIDE.Shell.Abstractions</RootNamespace>
+  </PropertyGroup>
+
+</Project>
