+++ src/AvaloniaIDE.Shell.Abstractions/IShellState.cs
@@ -1,6 +1,6 @@
 using System.Threading.Tasks;
 
-namespace AvaloniaIDE.Shell.State;
+namespace AvaloniaIDE.Shell.Abstractions;
 
 public interface IShellState
 {
