--- src/AvaloniaIDE.Shell.State/AvaloniaIDE.Shell.State.csproj
@@ -1,18 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <AssemblyName>AvaloniaIDE.Shell.State</AssemblyName>
-    <RootNamespace>AvaloniaIDE.Shell.State</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Hosting" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Serilog.Extensions.Hosting" />
-    <PackageReference Include="Serilog.Settings.Configuration" />
-    <PackageReference Include="Serilog.Sinks.File" />
-  </ItemGroup>
-
-</Project>
