+++ src/AvaloniaIDE.Shell.States/ShellStartedState.cs
@@ -1,6 +1,5 @@
-using AvaloniaIDE.Shell.State;
-using Avalonia;
-using Avalonia.Controls.ApplicationLifetimes;
+using AvaloniaIDE.Shell.Abstractions;
+using AvaloniaIDE.Shell.States.Logging;
 using System;
 using System.Threading.Tasks;
 using System.Threading;
@@ -8,8 +7,10 @@ using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Configuration;
+using Avalonia;
+using Avalonia.Controls.ApplicationLifetimes;
 
-namespace AvaloniaIDE.Shell.UI;
+namespace AvaloniaIDE.Shell.States;
 
 internal sealed class ShellStartedState : ShellStateBase
 {
