+++ src/AvaloniaIDE.Shell/AvaloniaIDE.Shell.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
         <OutputType>Exe</OutputType>
@@ -8,18 +8,18 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="Microsoft.Extensions.Configuration.Json" />
+        <ProjectReference Include="..\AvaloniaIDE.Shell.States\AvaloniaIDE.Shell.States.csproj" />
+        <ProjectReference Include="..\AvaloniaIDE.Shell.Abstractions\AvaloniaIDE.Shell.Abstractions.csproj" />
     </ItemGroup>
-
+    
     <ItemGroup>
-        <PackageReference Include="Serilog.Extensions.Hosting" />
-        <PackageReference Include="Serilog.Settings.Configuration" />
-        <PackageReference Include="Serilog.Sinks.File" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\AvaloniaIDE.Shell.State\AvaloniaIDE.Shell.State.csproj" />
-        <ProjectReference Include="..\AvaloniaIDE.Shell.Hosting\AvaloniaIDE.Shell.Hosting.csproj" />
+        <Content Include="Configuration/shellsettings.json" CopyToOutputDirectory="PreserveNewest"/>
+        <Content Condition="'$(Configuration)' == 'Debug'"
+                 Include="Configuration/shellsettings.Development.json"
+                 CopyToOutputDirectory="PreserveNewest"/>
+        <Content Condition="'$(Configuration)' == 'Release'"
+                 Include="Configuration/shellsettings.Production.json"
+                 CopyToOutputDirectory="PreserveNewest"/>
     </ItemGroup>
 
 </Project>
