+++ src/AvaloniaIDE.sln
@@ -5,13 +5,15 @@ VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AvaloniaIDE.Shell", "AvaloniaIDE.Shell\AvaloniaIDE.Shell.csproj", "{B8E15EA9-2A83-4E01-BAB2-0512B2D64761}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AvaloniaIDE.Shell.State", "AvaloniaIDE.Shell.State\AvaloniaIDE.Shell.State.csproj", "{FB57396D-8348-4020-9B75-3BE0BEFE04EE}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AvaloniaIDE.Shell.States", "AvaloniaIDE.Shell.States\AvaloniaIDE.Shell.States.csproj", "{FB57396D-8348-4020-9B75-3BE0BEFE04EE}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AvaloniaIDE.Shell.UI", "AvaloniaIDE.Shell.UI\AvaloniaIDE.Shell.UI.csproj", "{F613AAFB-378B-4F4A-9AD2-64E00B520188}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AvaloniaIDE.Shell.Hosting", "AvaloniaIDE.Shell.Hosting\AvaloniaIDE.Shell.Hosting.csproj", "{647B95A2-2836-4C9E-B330-6457D1195CE4}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AvaloniaIDE.AutomationClient", "AvaloniaIDE.AutomationClient\AvaloniaIDE.AutomationClient.csproj", "{5D746AC3-486F-47CE-960F-17D7F7B24E52}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AvaloniaIDE.Shell.Abstractions", "AvaloniaIDE.Shell.Abstractions\AvaloniaIDE.Shell.Abstractions.csproj", "{3070AA14-DB98-461B-B717-B7B860DCD041}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -42,5 +44,9 @@ Global
 		{5D746AC3-486F-47CE-960F-17D7F7B24E52}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5D746AC3-486F-47CE-960F-17D7F7B24E52}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5D746AC3-486F-47CE-960F-17D7F7B24E52}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3070AA14-DB98-461B-B717-B7B860DCD041}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3070AA14-DB98-461B-B717-B7B860DCD041}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3070AA14-DB98-461B-B717-B7B860DCD041}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3070AA14-DB98-461B-B717-B7B860DCD041}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
